add support for YAML version 1 to cleaning and parsing functions
Now 'widgets' is a *list* with keys 'name' and 'data', as opposed
to a mapping from 'name' to the contents of 'data'. This change
allows us to get a unique ordering for the widgets that is robust
to changes in the widget names, which allows us to safely re-upload
yaml without clobbering entries already in the database.