change the behaviour of the previous submission button
The previous behaviour was not particularly useful, as a common
workflow is to click "next ungraded", then want to go back to
the previous seen submission to correct it. Now we maintain a
stack of submission IDs that we have seen, and "previous" just
pops from this stack. An empty stack results in no change to the
submission.