Add bottom bar to PDF and remove margin parameter
The margin parameter is removed to make things simpler for the corner
marker detection. The bottom bar is added to be able to identify the
orientation of a page without depending on the barcode recognition.

Note: Git seems to think two of the changes constitute a rename, but I
actually deleted one file and created another.