Add (generated) exam code
Since the code now identifies the exam on barcodes (e.g. DataMatrices),
it is not necessary for the name to be unique anymore. To allow this,
data storage is now done using the exam ID instead of its name.

An advantage of this is that it can't break anymore if the exam name
contains a character which is not allowed by the filesystem.